Updated the LaTeX template.
Paragraphs are seperated with whitespace.
Paragraphs doesn't start with one line indentations.
Linenumbering is correctly setup with roman letters in the beginning and
numbers starting from Introduction chapter.
Table of content is added.
List of figures is added and shows up in the table of content.
List of tables is added and shows up in the table of content.